Move the function for retrieving Azure OpenAI API version for testing to test_utils.py
