+++ python/tests/integration/completions/test_utils.py
@@ -1,6 +1,7 @@
 # Copyright (c) Microsoft. All rights reserved.
 
 import logging
+import os
 import time
 
 logging.basicConfig(level=logging.DEBUG)
@@ -21,3 +22,27 @@ async def retry(func, retries=20):
             if i == retries - 1:  # Last retry
                 raise
             time.sleep(max(min(i, max_delay), min_delay))
+
+
+def get_aoai_api_versions():
+    """Retrieves a list of Azure OpenAI API versions for text completions
+
+    As Azure OpenAI may change the response content with different API versions,
+    it is necessary to test across multiple versions.
+    Supported API versions:
+        https://learn.microsoft.com/en-us/azure/ai-services/openai/reference#completions
+    """
+    api_versiosn = os.environ["AzureOpenAI__APIVersions"]
+    return api_versiosn.split(",")
+
+
+def get_aoai_chat_api_versions():
+    """Retrieves a list of Azure OpenAI API versions for chat completions
+
+    As Azure OpenAI may change the response content with different API versions,
+    it is necessary to test across multiple versions.
+    Supported API versions:
+        https://learn.microsoft.com/en-us/azure/ai-services/openai/reference#chat-completions
+    """
+    api_versions = os.environ["AzureOpenAIChat__APIVersions"]
+    return api_versions.split(",")
